distinguish between functions and methods, make join a function
